+++ src/apis/index.ts
@@ -1,7 +1,0 @@
-export * from './CountryApi';
-export * from './CurrencyApi';
-export * from './DomainApi';
-export * from './EmailApi';
-export * from './IPApi';
-export * from './PhoneApi';
-export * from './UtilsApi';