+++ src/index.ts
@@ -1,3 +1,10 @@
-export * from './runtime';
-export * from './apis';
-export * from './models';
+/**
+ * DEV.ME API Documentation
+ * DEV.ME API Documentation [Currency Conversion and Exchange Rates API](https://dev.me/products/currency) - [IP2Location, IP Country, IP Information API](https://dev.me/products/ip) -  [Email Validation, Mailbox Verification](https://dev.me/products/email) - [Phone Number Validation](https://dev.me/products/phone). You can learn more at [dev.me](https://dev.me). For this example you can use api key `demo-key` to test the APIs
+ *
+ * The version of the OpenAPI document: 1.0.0
+ * Contact: support@dev.me
+ */
+
+export * from './api';
+export * from './configuration';