+++ test/currency-test.ts
@@ -3,25 +3,25 @@ import { expect } from 'chai';
 
 describe('currency', async function () {
   it('should convert currency USD - EUR', async function () {
-    const result = await currencyApi.v1ConvertCurrency({
+    const { data } = await currencyApi.v1ConvertCurrency({
       amount: 10,
       from: 'USD',
       to: 'EUR',
     });
 
-    expect(result.from).to.be.eql('USD');
-    expect(result.to).to.be.eql('EUR');
+    expect(data.from).to.be.eql('USD');
+    expect(data.to).to.be.eql('EUR');
 
-    expect(result.convertedAmount).to.be.lessThan(9);
-    expect(result.convertedAmount).to.be.greaterThan(8);
+    expect(data.convertedAmount).to.be.lessThan(9);
+    expect(data.convertedAmount).to.be.greaterThan(8);
 
-    expect(result.exchangeRate).to.be.lessThan(0.9);
-    expect(result.exchangeRate).to.be.greaterThan(0.8);
+    expect(data.exchangeRate).to.be.lessThan(0.9);
+    expect(data.exchangeRate).to.be.greaterThan(0.8);
 
-    expect(result.originalAmount).to.be.eql(10);
+    expect(data.originalAmount).to.be.eql(10);
 
-    expect(result.convertedText.startsWith('10 USD equal to')).to.be.eql(true);
-    expect(result.rateTime.startsWith('2022-')).to.be.eql(true);
+    expect(data.convertedText.startsWith('10 USD equal to')).to.be.eql(true);
+    expect(data.rateTime.startsWith('2022-')).to.be.eql(true);
 
     // expect(result).to.be.eql({
     //   convertedAmount: 8.819,
@@ -34,8 +34,8 @@ describe('currency', async function () {
     // });
   });
   it('should list currencies', async function () {
-    const result = await currencyApi.v1ListCurrencies({ code: ['EUR', 'USD'] });
-    expect(result.list.length).to.be.eql(2);
-    expect(result.list[1].code).to.be.eql('USD');
+    const { data } = await currencyApi.v1ListCurrencies({ code: ['EUR', 'USD'] });
+    expect(data.list.length).to.be.eql(2);
+    expect(data.list[1].code).to.be.eql('USD');
   });
 });